rebase onto fully merged pedant
