+++ lib/pedant/rspec/search_util.rb
@@ -103,6 +103,23 @@ module Pedant
           end
         end
 
+        def can_perform_a_search_that_is_acl_filtered_for(object_type)
+          valid_object_type?(object_type)
+          it "should return filtered results when ACL on #{object_type}s exist" do
+            restrict_permissions_to "/#{object_type}s/#{base_object_name}_3",
+                                    normal_user => ["delete"]
+
+            # A little bit of confirmation that the ACL has applied correctly
+            n = get(api_url("/#{object_type}s/#{base_object_name}_3"), normal_user)
+            n.should look_like({:status => 403})
+
+            with_search_polling do
+              r = get("#{request_url}/?q=name:*", normal_user)
+              parse(r)["rows"].any? {|row| row["name"] == "#{base_object_name}_3"}.should be false
+            end
+          end
+        end
+
         def perform_a_search_that_returns_no_results(object_type)
           valid_object_type?(object_type)
           context 'a search that should return no results' do
@@ -433,6 +450,20 @@ module Pedant
             performing_a_search "should succeed, and return multiple #{object_type}s"
           end
         end # can_perform_basic_partial_search_for
+
+        def can_perform_a_partial_search_that_is_filtered_for(object_type)
+          valid_object_type? object_type
+          it "should return filtered results when ACLs exist" do
+            restrict_permissions_to "/#{object_type}s/#{base_object_name}_3",
+                                    normal_user => ["delete"]
+
+            payload = { "name" => ["name"] }
+            with_search_polling do
+              r = post("#{request_url}?q=name:*", normal_user, {payload: payload})
+              parse(r)["rows"].any? {|row| row["data"]["name"] == "#{base_object_name}_3"}.should be false
+            end
+          end
+        end
 
         # Helper method to determine the "real" search path that
         # should be submitted in the partial search request body.
